requirements.txt
pandas
streamlit-lottie
